% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{cluster_swarm}
\alias{cluster_swarm}
\title{Cluster Particles in Swarm}
\usage{
cluster_swarm(x.p, num_cluster, min_points)
}
\arguments{
\item{x.p}{positions of particles in swarm}

\item{num_cluster}{how many clusters should the particles be grouped into?}

\item{min_points}{what is the minimum number of partciles that should be in a cluster?}
}
\value{
cluster assignment for each particle in swarm

cluster centers
}
\description{
Group particles in swarm into clusters
}
