% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{convert_search_to_swarm}
\alias{convert_search_to_swarm}
\title{Convert Search Results into Initial Scedastic Surrogate Swarm Position}
\usage{
convert_search_to_swarm(
  search_state,
  usevec,
  desired_values,
  param_len,
  lowlim,
  highlim,
  ineq_w,
  eq_w,
  config = swarm.config()
)
}
\arguments{
\item{search_state}{search state list}

\item{usevec}{which of the search instances to use}

\item{param_len}{dimensionality of optimization problem}

\item{lowlim}{lower bound of input values}

\item{highlim}{upper bound of input values}

\item{ineq_w}{inequality constraint weights}

\item{eq_w}{equality contstraint weights}

\item{config}{swarm configuration list}

\item{desired_vals}{desired outputs}
}
\value{
swarm state list
}
\description{
Function to convert the results of a surrogate search into the initial locations
of a Scedastic Surrogate Swarm optimization, preserving the learned behaviour
by passing saved surrogates
}
