% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{error_function}
\alias{error_function}
\title{Customization of error function}
\usage{
error_function(actual, predicted, type = "abs")
}
\arguments{
\item{actual}{known values}

\item{predicted}{desired values}

\item{type}{what error function should be used?
\itemize{
\item{"abs"}{ Sum of absolute errors}
\item{"rme"}{ Root mean squared error}
}}
}
\value{
error value
}
\description{
Allows for different error functions to be used
}
