% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{fit_poly}
\alias{fit_poly}
\title{Fit Polynomial Instance}
\usage{
fit_poly(X_poly, y, crossval = F)
}
\arguments{
\item{X_poly}{Polynomial input array}

\item{y}{Response values to fit polynomial to}

\item{crossval}{Should cross validation be used in fitting?}
}
\value{
polynomial coefficients
}
\description{
Function to fit polynomial to observed data
}
