% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{fitness_fun}
\alias{fitness_fun}
\title{Fitness Function}
\usage{
fitness_fun(mod_vals, desired_vals, ineq_w, eq_w)
}
\arguments{
\item{mod_vals}{outputs of model function with current inputs}

\item{desired_vals}{desired outputs}

\item{ineq_w}{weights of inequality constraints}

\item{eq_w}{weights of equality constraints}
}
\value{
fitness score
}
\description{
Evaluate accuracy of solution
}
