% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{hetloss}
\alias{hetloss}
\title{Heteroscedastic Loss Function}
\usage{
hetloss(goal, mean, sd)
}
\arguments{
\item{goal}{desired outputs}

\item{mean}{modeled means}

\item{sd}{modeled standard deviations}
}
\value{
heteroscedastic loss
}
\description{
Function to determine heteroscedastic loss
}
