% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{initialize_search}
\alias{initialize_search}
\title{Initialize Surrogate Search Optimization}
\usage{
initialize_search(param_len, lowlim, highlim, config = search.config())
}
\arguments{
\item{param_len}{dimensionality of optimization problem}

\item{lowlim}{lower bound of input values}

\item{highlim}{upper bound of input values}

\item{config}{search configuration list}
}
\value{
search state list
}
\description{
Function to establish initial positions of surrogate searcher
}
