% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{initialize_swarm}
\alias{initialize_swarm}
\title{Initialize Scedastic Surrogate Swarm Optimization}
\usage{
initialize_swarm(
  desired_values,
  param_len,
  lowlim,
  highlim,
  ineq_w,
  eq_w,
  config = swarm.config()
)
}
\arguments{
\item{param_len}{dimensionality of optimization problem}

\item{lowlim}{lower bound of input values}

\item{highlim}{upper bound of input values}

\item{ineq_w}{inequality constraint weights}

\item{eq_w}{equality contstraint weights}

\item{config}{swarm configuration list}

\item{desired_vals}{desired outputs}
}
\value{
swarm state list
}
\description{
Function to establish initial positions and velocities of Scedastic Surrogate Swarm
}
