% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{make_poly}
\alias{make_poly}
\title{Construct Polynomial Instance}
\usage{
make_poly(dat, degree)
}
\arguments{
\item{dat}{input data}

\item{degree}{polynomial degree}
}
\value{
polynomial input array
}
\description{
This function creates a polynomial matrix of the specified degree provided
input data
}
