% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{modf}
\alias{modf}
\title{Model Function}
\usage{
modf(params)
}
\arguments{
\item{params}{model input array}
}
\value{
model output array with constraint values
}
\description{
Concatenate outputs of user defined model function with user defined inequality
and equality constraints
}
