% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{normalize}
\alias{normalize}
\title{Generalized Normalization Function}
\usage{
normalize(x)
}
\arguments{
\item{x}{unnormalized array}
}
\value{
normalized array
}
\description{
Function to perform robust normalization in case the sum of x is 0.
}
