% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{poly_obj_fn}
\alias{poly_obj_fn}
\title{Objective Function for Suboptimization Problem}
\usage{
poly_obj_fn(xsel, polys, desired_values, centers, deg)
}
\arguments{
\item{xsel}{selected input values to evaluate surrogate for}

\item{polys}{all polynomials fit to explored responses}

\item{centers}{polynomial centers}

\item{deg}{degree of polynomial (recommended between 0 and 1)}

\item{desired_vals}{desired outputs}
}
\value{
objective function score
}
\description{
Construct objective function for the suboptimization
}
