% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{sample_lhs_around_center}
\alias{sample_lhs_around_center}
\title{Perform Latin Hypercube Sampling around a Center Point}
\usage{
sample_lhs_around_center(center, n_samples, range_min, range_max)
}
\arguments{
\item{center}{center point location}

\item{n_samples}{number of Latin Hypercube samples to perform}

\item{range_min}{lower bound of sampling}

\item{range_max}{upper bound of sampling}
}
\value{
latin hypercube samples
}
\description{
Function to do Latin Hypercube Sampling around a central location based on
allowable ranges
}
