% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{search.config}
\alias{search.config}
\title{Search Function configuration}
\usage{
search.config(deg = 2, gen = F, search_mag = NULL, search_samples = NULL)
}
\arguments{
\item{deg}{degree of surrogate polynomial}

\item{gen}{should a new search be initiated from a new location}

\item{search_mag}{magnitude of the local search}

\item{search_samples}{number of searches to perform around sample point}
}
\value{
search configuration parameter list
}
\description{
Establish parameters for surrogate search optimization
}
