% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{step_search}
\alias{step_search}
\title{Step through surrogate search optimization}
\usage{
step_search(
  desired_values,
  search_state,
  ineq_w,
  eq_w,
  config = search.config()
)
}
\arguments{
\item{search_state}{search state list}

\item{ineq_w}{inequality constraint weights}

\item{eq_w}{equality constraint weights}

\item{config}{search configuration list}

\item{desired_vals}{desired outputs}
}
\value{
search state list
}
\description{
Function to step through surrogate search optimization by updating search states
}
