% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{step_swarm}
\alias{step_swarm}
\title{Step through Scedastic Surrogate Swarm optimization}
\usage{
step_swarm(desired_values, swarm_state, ineq_w, eq_w, config = swarm.config())
}
\arguments{
\item{swarm_state}{swarm state list}

\item{ineq_w}{inequality constraint weights}

\item{eq_w}{equality constraint weights}

\item{config}{swarm configuration list}

\item{desired_vals}{desired outputs}
}
\value{
swarm state list
}
\description{
Function to step through Scedastic Surrogate Swarm optimization by updating swarm states
}
