% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{surrogate_model}
\alias{surrogate_model}
\title{Evaluate Surrogate Model}
\usage{
surrogate_model(xsel, polys, centers, deg, type = "closest", form = "matrix")
}
\arguments{
\item{xsel}{selected input values to evaluate surrogate for}

\item{polys}{all polynomials fit to explored responses}

\item{centers}{polynomial centers}

\item{deg}{degree of polynomial (recommended between 0 and 1)}

\item{type}{should we only use the polynomial evaluated with the center closest to xsel?
\itemize{
\item{"closest"}{ Only use the polynomial whose center is closest to xsel}
\item{"all"}{ Evaluate all polynomials and do weighting average based on distance of respective centers to xsel}
}}

\item{form}{should distance be evaluated with a simple absolute distance?
\itemize{
\item{"matrix"}{ Evaluate distance as sum of absolute differences}
\item{"errorfun"}{ Evaluate distance with error function}
}}
}
\value{
predicted model outputs based on surrogate
}
\description{
Evaluate the constructed surrogate model of the true function based on proximity
to centers and respective polynomials fit to response
}
