% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{swarm.config}
\alias{swarm.config}
\title{Scedastic Surrogate Swarm Configuration}
\usage{
swarm.config(
  swarm_size = 100,
  num_cluster = 1,
  min_points = 3,
  deg = 1,
  localityfac = 0.6,
  opt_from_best = F,
  swarm.control = list(stoch_w = 0.1, vel_w = 0.5, r_p_w = 0.3, r_g_w = 0.6, r_go_w =
    0.6, poly_w = 0.2, sens_overall = F)
)
}
\arguments{
\item{swarm_size}{number of particles in swarm}

\item{num_cluster}{number of clusters for grouping particles}

\item{min_points}{minimum number of particles per cluster}

\item{deg}{degree of surrogate polynomial}

\item{localityfac}{what level of locality to use when determining global values (1 would be full global topology)}

\item{opt_from_best}{when performing suboptimization should we start from the best known position?}

\item{swarm.control}{swarm control weight list}
}
\value{
swarm configuration parameter list
}
\description{
Establish parameters for evaluating Scedastic Surrogate Swarm
}
