% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_functions.R
\name{swarm.control}
\alias{swarm.control}
\title{Scedastic Surrogate Swarm Control}
\usage{
swarm.control(
  stoch_w = 0.1,
  vel_w = 0.5,
  r_p_w = 0.3,
  r_g_w = 0.6,
  r_go_w = 0.6,
  poly_w = 0.2,
  sens_overall = F
)
}
\arguments{
\item{stoch_w}{how much influence does scedastic factor have?}

\item{vel_w}{how much influence does the previous velocity have?}

\item{r_p_w}{how much influence does the particle's historical best position have?}

\item{r_g_w}{how much influence does the global current best position have?}

\item{r_go_w}{how much influence does the global historical best position have?}

\item{poly_w}{how much influence does the suboptimized surrogate function have?}

\item{sens_overall}{do we use the overall sensitivity when backpropagating the stochastic factor?}
}
\value{
control list
}
\description{
Establish weights and importances for Scedastic Surrogate Swarm for determining new
particle velocities.
}
